change unsigned int settings to uint
